requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Werkzeug==2.3.7
requests==2.31.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
PyYAML==6.0.1
python-dotenv==1.0.0
stripe==7.7.0
itsdangerous==2.2.0
